feat: moved MixerConnectionTypes to MixerProtocolInterface.ts
