+++ shared/src/constants/mixerProtocols/genericMidi.ts
@@ -1,5 +1,4 @@
-import { IMixerProtocol, emptyMixerMessage } from '../MixerProtocolInterface'
-import { MixerConnectionTypes } from '../MixerProtocolPresets'
+import { IMixerProtocol, emptyMixerMessage, MixerConnectionTypes } from '../MixerProtocolInterface'
 
 export const GenericMidi: IMixerProtocol = {
     protocol: MixerConnectionTypes.GenericMidi,
